backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
groq==1.2.0
python-dotenv==1.0.1
pillow==10.4.0
pydantic==2.9.2
motor==3.6.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.2.0
